CMFM: Make sure no NPE happens
Fix for the case when getExternalFilesDir returns null

+++ src/com/cyanogenmod/filemanager/console/secure/SecureConsole.java
@@ -624,15 +624,17 @@ public class SecureConsole extends VirtualMountPointConsole {
      */
     private void clearCache(Context ctx) {
         File filesDir = ctx.getExternalFilesDir(null);
-        File[] cacheFiles = filesDir.listFiles(new FilenameFilter() {
-            @Override
-            public boolean accept(File dir, String filename) {
-                return filename.startsWith(mStorageName)
-                        && filename.endsWith(".tmp");
+        if (filesDir != null) {
+            File[] cacheFiles = filesDir.listFiles(new FilenameFilter() {
+                @Override
+                public boolean accept(File dir, String filename) {
+                    return filename.startsWith(mStorageName)
+                            && filename.endsWith(".tmp");
+                }
+            });
+            for (File cacheFile : cacheFiles) {
+                cacheFile.delete();
             }
-        });
-        for (File cacheFile : cacheFiles) {
-            cacheFile.delete();
         }
     }
 }
